pyproject.toml
[tool.poetry]
name = "tabular-models-benchmark"
version = "0.1.0"
description = ""
authors = ["victorlandeau <vlandeau@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
jupyter = "^1.0.0"
sklearn = "^0.0"
pandas = "^1.2.1"
plotly = "^4.14.3"
jedi = "0.17.2"
xgboost = "^1.3.3"
catboost = "^0.24.4"
lightgbm = "^3.1.1"
optuna = "^2.4.0"
PyHamcrest = "^2.0.2"
pytest = "^6.2.2"
category-encoders = "^2.2.2"
jupytext = "^1.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
